Tidy tests
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/validator/trunk@1094808 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/validator/routines/BigDecimalValidatorTest.java
@@ -120,8 +120,8 @@ public class BigDecimalValidatorTest extends AbstractNumberValidatorTest {
         BigDecimal number20 = new BigDecimal("20");
         BigDecimal number21 = new BigDecimal("21");
         
-        float min = (float)10;
-        float max = (float)20;
+        float min = 10;
+        float max = 20;
 
         // Test isInRange()
         assertFalse("isInRange(A) < min",   validator.isInRange(number9,  min, max));
